avx512: add macros to declare mask(z) functions, start using them
I haven't decided if I want to proceed with this or not, but I'm
leaning towards it.  I've already run into several cases where they
can't be used, but in those cases we can fall back on writing out
the functions.  That makes the code for those cases a bit different,
but I think it's manageable.